requirements-dev.txt
wheel
flake8
flake8-colors
flake8-docstrings
requests
inflection
twine
setuptools
PyYAML
bumpversion
pytest
pytest-vcr
coverage
